+++ neunet/nn/layers/embedding.py
@@ -23,7 +23,7 @@ class Embedding():
         self.embedding_dim = embedding_dim
 
         stdv = 1. / np.sqrt(embedding_dim)
-        self.weight = Tensor(np.random.uniform(-stdv, stdv, (num_embeddings, embedding_dim)))
+        self.weight = Tensor(np.random.uniform(-stdv, stdv, (num_embeddings, embedding_dim)), dtype=np.float32)
 
     def one_hot(self, X):
         O = np.zeros((X.size, self.num_embeddings))
