+++ neunet/nn/layers/gru.py
@@ -120,18 +120,18 @@ class GRU():
 
 
         stdv = 1. / np.sqrt(self.hidden_size)
-        self.weight_z = Tensor(np.random.uniform(-stdv, stdv, (self.input_size, self.hidden_size)))
-        self.weight_r = Tensor(np.random.uniform(-stdv, stdv, (self.input_size, self.hidden_size)))
-        self.weight_h = Tensor(np.random.uniform(-stdv, stdv, (self.input_size, self.hidden_size)))
+        self.weight_z = Tensor(np.random.uniform(-stdv, stdv, (self.input_size, self.hidden_size)), dtype=np.float32)
+        self.weight_r = Tensor(np.random.uniform(-stdv, stdv, (self.input_size, self.hidden_size)), dtype=np.float32)
+        self.weight_h = Tensor(np.random.uniform(-stdv, stdv, (self.input_size, self.hidden_size)), dtype=np.float32)
 
-        self.weight_hz = Tensor(np.random.uniform(-stdv, stdv, (self.hidden_size, self.hidden_size)))
-        self.weight_hr = Tensor(np.random.uniform(-stdv, stdv, (self.hidden_size, self.hidden_size)))
-        self.weight_hh = Tensor(np.random.uniform(-stdv, stdv, (self.hidden_size, self.hidden_size)))
+        self.weight_hz = Tensor(np.random.uniform(-stdv, stdv, (self.hidden_size, self.hidden_size)), dtype=np.float32)
+        self.weight_hr = Tensor(np.random.uniform(-stdv, stdv, (self.hidden_size, self.hidden_size)), dtype=np.float32)
+        self.weight_hh = Tensor(np.random.uniform(-stdv, stdv, (self.hidden_size, self.hidden_size)), dtype=np.float32)
 
         if bias:
-            self.bias_z = Tensor(np.zeros(self.hidden_size))
-            self.bias_r = Tensor(np.zeros(self.hidden_size))
-            self.bias_h = Tensor(np.zeros(self.hidden_size))
+            self.bias_z = Tensor(np.zeros(self.hidden_size), dtype=np.float32)
+            self.bias_r = Tensor(np.zeros(self.hidden_size), dtype=np.float32)
+            self.bias_h = Tensor(np.zeros(self.hidden_size), dtype=np.float32)
         else:
             self.bias_z = None
             self.bias_r = None
